externalise xls reader
